+++ core/src/test/java/com/graphhopper/search/EdgeKVStorageTest.java
@@ -10,15 +10,15 @@ import java.io.File;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 
-import static com.graphhopper.search.StringIndex.MAX_UNIQUE_KEYS;
+import static com.graphhopper.search.EdgeKVStorage.MAX_UNIQUE_KEYS;
 import static org.junit.jupiter.api.Assertions.*;
 
-public class StringIndexTest {
+public class EdgeKVStorageTest {
 
-    private final static String location = "./target/stringindex-store";
+    private final static String location = "./target/edge-kv-storage";
 
-    private StringIndex create() {
-        return new StringIndex(new RAMDirectory(), 1000).create(1000);
+    private EdgeKVStorage create() {
+        return new EdgeKVStorage(new RAMDirectory(), 1000).create(1000);
     }
 
     Map<String, Object> createMap(String... strings) {
@@ -33,7 +33,7 @@ public class StringIndexTest {
 
     @Test
     public void putSame() {
-        StringIndex index = create();
+        EdgeKVStorage index = create();
         long aPointer = index.add(createMap("a", "same name", "b", "same name"));
 
         assertNull(index.get(aPointer, ""));
@@ -48,7 +48,7 @@ public class StringIndexTest {
 
     @Test
     public void putAB() {
-        StringIndex index = create();
+        EdgeKVStorage index = create();
         long aPointer = index.add(createMap("a", "a name", "b", "b name"));
 
         assertNull(index.get(aPointer, ""));
@@ -58,7 +58,7 @@ public class StringIndexTest {
 
     @Test
     public void putEmpty() {
-        StringIndex index = create();
+        EdgeKVStorage index = create();
         assertEquals(1, index.add(createMap("", "")));
         assertEquals(5, index.add(createMap("", null)));
         // cannot store null value if it is the first value of the key:
@@ -72,7 +72,7 @@ public class StringIndexTest {
 
     @Test
     public void putMany() {
-        StringIndex index = create();
+        EdgeKVStorage index = create();
         long aPointer = 0, tmpPointer = 0;
 
         for (int i = 0; i < 10000; i++) {
@@ -91,7 +91,7 @@ public class StringIndexTest {
 
     @Test
     public void putManyKeys() {
-        StringIndex index = create();
+        EdgeKVStorage index = create();
         // one key is already stored => empty key
         for (int i = 1; i < MAX_UNIQUE_KEYS; i++) {
             index.add(createMap("a" + i, "a name"));
@@ -105,7 +105,7 @@ public class StringIndexTest {
 
     @Test
     public void putDuplicate() {
-        StringIndex index = create();
+        EdgeKVStorage index = create();
         long aPointer = index.add(createMap("a", "longer name", "b", "longer name"));
         long bPointer = index.add(createMap("c", "longer other name"));
         // value storage: 1 byte for count, 2 bytes for keyIndex and 4 bytes for delta of dup_marker and 3 bytes (keyIndex + length for "longer name")
@@ -130,7 +130,7 @@ public class StringIndexTest {
 
     @Test
     public void testNoErrorOnLargeStringValue() {
-        StringIndex index = create();
+        EdgeKVStorage index = create();
         String str = "";
         for (int i = 0; i < 127; i++) {
             str += "ß";
@@ -142,7 +142,7 @@ public class StringIndexTest {
 
     @Test
     public void testTooLongStringValueError() {
-        StringIndex index = create();
+        EdgeKVStorage index = create();
         assertThrows(IllegalArgumentException.class, () -> index.add(createMap("", "Бухарестская улица (http://ru.wikipedia.org/wiki" +
                 "/%D0%91%D1%83%D1%85%D0%B0%D1%80%D0%B5%D1%81%D1%82%D1%81%D0%BA%D0%B0%D1%8F_%D1%83%D0%BB%D0%B8%D1%86%D0%B0_(%D0%A1%D0%B0%D0%BD%D0%BA%D1%82-%D0%9F%D0%B5%D1%82%D0%B5%D1%80%D0%B1%D1%83%D1%80%D0%B3))")));
 
@@ -160,7 +160,7 @@ public class StringIndexTest {
 
     @Test
     public void testNoErrorOnLargestByteArray() {
-        StringIndex index = create();
+        EdgeKVStorage index = create();
         byte[] bytes = new byte[255];
         byte[] copy = new byte[255];
         for (int i = 0; i < bytes.length; i++) {
@@ -178,7 +178,7 @@ public class StringIndexTest {
 
     @Test
     public void testIntLongDoubleFloat() {
-        StringIndex index = create();
+        EdgeKVStorage index = create();
         long intres = index.add(Collections.singletonMap("intres", 4));
         long doubleres = index.add(Collections.singletonMap("doubleres", 4d));
         long floatres = index.add(Collections.singletonMap("floatres", 4f));
@@ -196,7 +196,7 @@ public class StringIndexTest {
 
     @Test
     public void testIntLongDoubleFloat2() {
-        StringIndex index = create();
+        EdgeKVStorage index = create();
         Map<String, Object> map = new HashMap<>();
         map.put("int", 4);
         map.put("long", 4L);
@@ -220,12 +220,12 @@ public class StringIndexTest {
     public void testFlush() {
         Helper.removeDir(new File(location));
 
-        StringIndex index = new StringIndex(new RAMDirectory(location, true).create(), 1000);
+        EdgeKVStorage index = new EdgeKVStorage(new RAMDirectory(location, true).create(), 1000);
         long pointer = index.add(createMap("", "test"));
         index.flush();
         index.close();
 
-        index = new StringIndex(new RAMDirectory(location, true), 1000);
+        index = new EdgeKVStorage(new RAMDirectory(location, true), 1000);
         assertTrue(index.loadExisting());
         assertEquals("test", index.get(pointer, ""));
         // make sure bytePointer is correctly set after loadExisting
@@ -240,7 +240,7 @@ public class StringIndexTest {
     public void testLoadKeys() {
         Helper.removeDir(new File(location));
 
-        StringIndex index = new StringIndex(new RAMDirectory(location, true).create(), 1000).create(1000);
+        EdgeKVStorage index = new EdgeKVStorage(new RAMDirectory(location, true).create(), 1000).create(1000);
         long pointerA = index.add(createMap("c", "test value"));
         assertEquals(2, index.getKeys().size());
         long pointerB = index.add(createMap("a", "value", "b", "another value"));
@@ -249,7 +249,7 @@ public class StringIndexTest {
         index.flush();
         index.close();
 
-        index = new StringIndex(new RAMDirectory(location, true), 1000);
+        index = new EdgeKVStorage(new RAMDirectory(location, true), 1000);
         assertTrue(index.loadExisting());
         assertEquals("[, c, a, b]", index.getKeys().toString());
         assertEquals("test value", index.get(pointerA, "c"));
@@ -266,7 +266,7 @@ public class StringIndexTest {
 
     @Test
     public void testEmptyKey() {
-        StringIndex index = create();
+        EdgeKVStorage index = create();
         long pointerA = index.add(createMap("", "test value"));
         long pointerB = index.add(createMap("a", "value", "b", "another value"));
 
@@ -281,7 +281,7 @@ public class StringIndexTest {
     public void testRandom() {
         final long seed = new Random().nextLong();
         try {
-            StringIndex index = new StringIndex(new RAMDirectory(location, true).create()).create(1000);
+            EdgeKVStorage index = new EdgeKVStorage(new RAMDirectory(location, true).create()).create(1000);
             Random random = new Random(seed);
             List<String> keys = createRandomStringList(random, "_key", 100);
             List<Integer> values = createRandomList(random, 500);
@@ -310,7 +310,7 @@ public class StringIndexTest {
             index.flush();
             index.close();
 
-            index = new StringIndex(new RAMDirectory(location, true).create());
+            index = new EdgeKVStorage(new RAMDirectory(location, true).create());
             assertTrue(index.loadExisting());
             for (int i = 0; i < size; i++) {
                 Map<String, Object> map = index.getAll(pointers.get(i));